# Use ensure_dict in _optimize_insert_futures